[Bronze V] Title: 이상한 기호, Time: 212 ms, Memory: 17672 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) throws IOException {
         Scanner sc = new Scanner(System.in);
-        int userA = sc.nextInt();
-        int userB = sc.nextInt();
+        long userA = sc.nextInt();
+        long userB = sc.nextInt();
 
         System.out.println( (userA + userB) * (userA-userB) );
     }
